backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
anthropic==0.28.0
pydantic==2.7.1
numpy==1.26.4
scipy==1.13.1
python-multipart==0.0.9
httpx==0.27.0
